+++ src/routes/[taxonomy]/[term]/+page.svelte
@@ -1,5 +1,5 @@
 <script>
-	import NothingFound from '$lib/components/nothing-found.svelte';
+	import NoPosts from '$lib/components/no-posts.svelte';
 	import PostsList from '$lib/components/posts-list.svelte';
 	import Seo from 'svelte-seo';
 
@@ -20,6 +20,6 @@
 	{#if data.posts?.length}
 		<PostsList class="recent-posts" posts={data.posts} />
 	{:else}
-		<NothingFound />
+		<NoPosts />
 	{/if}
 </div>
