+++ src/routes/blog/+page.svelte
@@ -1,5 +1,5 @@
 <script>
-	import NothingFound from '$lib/components/nothing-found.svelte';
+	import NoPosts from '$lib/components/no-posts.svelte';
 	import PostsList from '$lib/components/posts-list.svelte';
 	import Seo from 'svelte-seo';
 
@@ -14,6 +14,6 @@
 		<h1>Recent Posts</h1>
 		<PostsList class="recent-posts" posts={data.posts} />
 	{:else}
-		<NothingFound />
+		<NoPosts />
 	{/if}
 </div>
